tests/wpt/web-platform-tests/screen-orientation/unlock.html

<!DOCTYPE html>
<meta charset="utf-8" />
<meta viewport="width=device-width, initial-scale=1" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script type="module">
  import {
    getOppositeOrientation,
    makeCleanup,
    attachIframe,
  } from "./resources/orientation-utils.js";

  test(() => {
    screen.orientation.unlock();
  }, "unlock() doesn't throw when there is no lock");

  test(() => {
    const value = screen.orientation.unlock();
    assert_equals(value, undefined);
  }, "unlock() returns a void value");

  promise_test(async (t) => {
    t.add_cleanup(makeCleanup());
    await test_driver.bless("request full screen");
    await document.documentElement.requestFullscreen();
    screen.orientation.unlock();
  }, "unlock() doesn't throw when there is no lock with fullscreen");

  promise_test(async (t) => {
    t.add_cleanup(makeCleanup());
    await test_driver.bless("request full screen");
    await document.documentElement.requestFullscreen();
    const promise = screen.orientation.lock(getOppositeOrientation());
    screen.orientation.unlock();
    await promise_rejects_dom(t, "AbortError", promise);
  }, "unlock() aborts a pending lock request");

  promise_test(async (t) => {
    t.add_cleanup(makeCleanup());
    await test_driver.bless("request full screen");
    await document.documentElement.requestFullscreen();
    const iframe = await attachIframe();
    const promise = iframe.contentWindow.screen.orientation.lock(
      getOppositeOrientation()
    );
    screen.orientation.unlock();
    await promise_rejects_dom(
      t,
      "AbortError",
      iframe.contentWindow.DOMException,
      promise
    );
  }, "unlock() aborts a pending lock request across documents");
</script>
